chore(Rng): Add prelude module
