+++ src/prelude.rs
@@ -1,0 +1,22 @@
+pub use turborand::{ForkableCore, GenCore, SecureCore, SeededCore, TurboCore, TurboRand};
+
+#[cfg(feature = "wyrand")]
+pub use turborand::prelude::Rng;
+
+#[cfg(feature = "chacha")]
+pub use turborand::prelude::ChaChaRng;
+
+#[cfg(feature = "rand")]
+pub use turborand::prelude::RandBorrowed;
+
+#[cfg(feature = "chacha")]
+pub use crate::component::chacha::ChaChaRngComponent;
+#[cfg(feature = "wyrand")]
+pub use crate::component::rng::RngComponent;
+#[cfg(feature = "chacha")]
+pub use crate::global::chacha::GlobalChaChaRng;
+#[cfg(feature = "wyrand")]
+pub use crate::global::rng::GlobalRng;
+#[cfg(any(feature = "wyrand", feature = "chacha"))]
+pub use crate::plugin::RngPlugin;
+pub use crate::traits::DelegatedRng;
